requirements.txt
requests>=2.31.0
psutil>=5.9.0
netifaces>=0.11.0
scapy>=2.5.0
dnspython>=2.4.0
speedtest-cli>=2.1.3
pillow>=10.0.0
customtkinter>=5.2.0
darkdetect>=0.8.0
